services/user-service/requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
pymongo==4.3.3
confluent-kafka==2.1.1
python-dotenv==1.0.0
uuid==1.30
python-dateutil==2.8.2
flask-cors==4.0.0
marshmallow==3.20.1